+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This file is used to list changes made in each version of the splunk cookbook.
 
 ## 4.0.2 (TBD)
 - Modifies the `chef-splunk::shclustering` to deploy a Splunk Search Head deployer
+- Fixes a regression made by commit 26fa04d9: when `node['splunk']['runasroot']` is false,
+  splunk isn't started with a non-root user or the startup scripts are not modified to allow
+  for non-root splunk commands
 
 ## 4.0.1 (2019-12-19)
 - Fixes [#130](https://github.com/chef-cookbooks/chef-splunk/issues/130)